cint: added rounding when converting to integers
TEST: verify that arguments are rounded up when their value is 0.5 or
higher.

Both GW-Basic and QBasic round the argument values.  Other conversions
to integer are also rounded up.  Therefore, the execute convert to
integer function also needed to round values, so was modified the call
the standard library round function to round the value on top of the
stack before checking for an overflow.