rnd: implemented compiling of no argument random function
TEST: verify that a RND function with no arguments is compiled to the
end of the expression.

There will be two forms of the RND function, one with no arguments that
returns a double random value from 0 to 1, and one with an integer
argument that returns an integer random value from 0 to the argument
specified.

The codes instance for the random codes has different requirements then
the existing function codes classes, so a new random function codes
class was added.  To start this class a single none code member, which
is a function code with a none argument type (a new none argument type
enumerator was added).  The argument data type function just returns a
null data type for now, but is not actually called yet.

The rnd code instance was added with null function pointers along with
the rnd codes instance with its keyword.  These changes were not enough
to get the test to pass because an open parenthesis character was still
being expected after the function keyword.

A new has arguments virtual function was added to the function codes
base class.  For the new random function codes class, this function
returns false since the rnd code does not have arguments.  So that not
all of the other function codes classes needed this function, the base
class function returns true.

In the expression compile class, the call to the add to the add function
code function was moved from the compile function arguments function
to the compile function function.  The compile function arguments
function was then modified to only check for an open parentheses and
call the compile parentheses function if the codes passed has arguments,
otherwise return a null data type (which won't be used).

As noted above, the random function codes argument data type is not
covered by any of the tests, but was required to be implemented since it
is defined as a pure virtual function in the base class.  It will be
called once the other RND function is implemented.