rnd: implemented compiler of integer argument random function
TEST: verify that a RND function with an integer argument is compiled to
the end of the expression.

This is the second form of the RND function.  The random function codes
class was given a new integer code member, a function code taking an
integer argument.  The constructor and code values functions were
updated for the new code member.

With the new RND code, the argument is now considered 'optional' so the
has arguments function no longer applies.  Therefore, this function was
replaced with an argument optional function, which returns false for the
function codes class and true for the random function codes class.

The compile function arguments function was modified to check the result
of this function after seeing the next character is not an open
parentheses.  If the argument is optional, then a null data type (which
is not used) is return, otherwise an "expected opening parentheses"
error is thrown.

The no arguments function will be needed for other functions that only
contain no arguments (for example, the DATE$ function), but it is not
needed now, so following the rule not to add features that are not
needed or used, this function was removed for now.