rnd: completed the compiling of the random function
TESTS: verify the each form of the RND function returns the correct data
type and the correct code is added to the program.

Each of the four tests were added one at a time and made to pass.  The
select function for the random function codes class needs to know the
number of arguments that were compiled (in this case zero or one
argument).  For convenience and to keep the number of arguments to the
select function minimal, the argument was changed from a single data
type to a vector of data types.  This will allow any number of argument
data types to be passed, and the number of arguments will be the size of
the vector.

The selection function of the mutli-type function codes class was
changed to check the data type of the first (front) of the vector.  The
random function codes select function checks if the vector is empty, and
then returning the none code and double data type, otherwise returning
the integer code and integer data type.  The other function codes
classes don't use this vector argument.

The compile function arguments function of the expression compiler was
changed to return a vector of data types, one data type if an argument
was compiled, or an empty vector if there were no arguments (and the
codes indicated that the argument was optional).  The name of the
variable receiving the return value from this function in the compile
function argument was renamed to reflect what is now returned.  The
argument of the add function code was changed to a vector of data types,
which is passed to the codes select function.