rnd: implemented recreation of random function with no arguments
TEST: verify that a RND function with no arguments is recreated
correctly.

Up to now, all functions had a single argument, but the RND function
with no arguments is the first exception (there will be more when the
string functions are implemented since they will have from zero to
three arguments).

The first attempt to get this test to pass required a bit of work to
give the recreate function access to the number of arguments the
function being recreated has.  A simple solution was found as the
commit message for those changes were being composed.

The simple solution was to add new recreate function with no arguments
functions.  The member function just needs to push the keyword of the
function code to the recreator stack.  The null pointer in the rnd code
was changed to a pointer to the new global function that calls the
recreator virtual member function.

Since the existing recreate function will no longer be a generic
function for handling the recreation of all functions, but will just
handle one argument functions, it was renamed to the recreate function
with one argument.  (There will be two and three argument versions when
the time comes.)