rnd: implemented initial random no argument execution function
TEST: verify that the RND no argument function is executed and returns a
number from 0 to 1.

The RND function will be difficult to test since the test cannot know
what random value it will produce (though will be predictable for a
given system, there is no  guarantee that another system or version of
the compiler environment will produce the same random value).

Therefore, to start (simple) this first test just verifies that 'a'
number from 0 to 1 exclusive is returned (0 is possible, but unlikely).
The test parses the output by putting it into a second input string
stream and reading the value.  It verifies that a number is parsed
correctly and that there is no extra output except for white space.

To get this test to pass, the execute rnd function just returns a
hard-coded value between 0 and 1.  This allowed the execute function to
be implemented (and its pointer replacing the null pointer in the rnd
code instance).  This also required a new push value function be added
to the executer class since there is no argument value on the stack that
can be replaced with the return value.  This push function was
implemented as template function since it will need to support integer
(and possible string) type later.