rnd: refactored test with a help function
The first test of the RND function was a bit clunky.  Some of the test
assertions (requires) and other statements were not directly related to
the tests (checking if the number is parsed, skipping white space,
checking for the correct end of the output).

The parse numbers template helper function was added to remove this
unrelated code from the test.  This function takes a reference to an
output string stream, places the string of this into an input string
stream, and proceeds to parse values (of the type passed as the template
argument), appends each to a vector (of the template argument type)
until there are no more (extract fails), and returns the vector of
values.  The caller checks if there is the proper end by checking the
size of the vector.  This was made a template function so that other
types can be used (the integer type will be used for the random with
integer argument function tests).

The test was than changed to call this new template function for the
double type, check if the size of the vector returned is one and
proceeds to verify the first (front) value of the vector.