rnd: changed execution to return actual random values
TEST: verify that actual unique random values are returned from the RND
function when executed.

Specific random numbers cannot be checked for even though generally will
be repeatable (even if not seeded since the random number generator is
seeded by default).  There is no guarantee that a different platform or
even different version of the compiler will give the same number.

Therefore, the test generates a set of random numbers, which is
accomplished by concatenating a set of print RND functions (by a
constant that is set to 10), and then executing these prints.  The
output is then parsed to a vector (using the previously implemented
parse values helper function).  The size of the vector return is checked
to make sure all the numbers were printed.  Each number is then added to
a set.  The final test verifies that the size of the set is the same as
the number of values generated (in other words, all of the numbers are
unique - it is unlikely out of a small set of random values that two
will be the same).  This test initially failed because the single
hard-coded random value was being returned.

The C++11 random facilities were used.  The default random engine was
used (which should be sufficient here), and a global instance of this
was added to the executer source file (the same generator should be used
throughout the application).  A uniform real distribution instance was
added to the executer class to generated values from zero (inclusive) to
one (exclusive).  Finally, a get random number function was added that
returns the value from calling the uniform distribution from the random
number generator.

The execute rnd function was changed from pushing a hard-coded value to
pushing the value returned from calling the get random number function.