rnd: implemented execution of random integer argument function
TEST: verify that the RND integer argument function is executed and
returns a numbers from 1 and the argument.

For a limited range of possible integer numbers, it is more likely that
a number will be repeated, so the same type of test used for the RND no
argument function will not work here.

Instead, a larger set of random numbers is drawn (from 1 to 10).  The
test verifies that all of the samples executed (50 samples are taken).
The random numbers vector is iterated.  The minimum and maximum numbers
of all of the samples is determined, and each number is added to a set.
Only unique numbers are added to the set.  Once complete, the number of
unique numbers (the size of the set) should be the same as the number of
possible random numbers (10).  The minimum number is checked for one,
and the maximum number is checked for ten.

Another get random number function was added to the executer class.
This function takes an integer argument, the limit, and returns an
integer.  Since the limit will vary, a single uniform integer
distribution instance was not added as a member, but only a local
temporary instance is created with one as the lower limit and the limit
argument as the upper limit of the distribution (inclusive).

The execute rnd integer function was added that gets the top integer
from the execution stack, calls the new get random number function with
this value and sets the top value of the execution stack.  A pointer to
this function replaces the null pointer in the random int code instance.