rnd: added argument error checks for random integer function
TESTS: verify that a run error is produced when the integer argument of
the RND function zero or negative.

Only non-zero positive arguments are valid with the random function with
an integer argument.  The execute random integer function was modified
to check the argument value and if less than or equal to zero, then a
run error is thrown.