project: added release notes for version 0.0.4
Added a functions integration test for the standalone IBC program
demonstrating the numeric functions.  This test does not add any new
coverage beyond the existing unit tests.

OTHER: removed a FIXME comment about the conversion of the double
constant value to and integer value, and while the conversion would lead
to a bogus integer value if the double is out of the range of an
integer, the integer value will not actually be used since there is a
check for valid range when a constant is changed to an integer constant.