const-num: round integer equivalent value of double constant
TEST: verify that the integer equivalent value of a double constant is
rounded correctly.

Numerical constants are stored as both a double value and an integer
value.  This eliminates extra conversions from one data type to another
while the program is running.  An integer constant (no decimal point, no
exponent, and within an integer value range) is also stored as a double.

A double constant is also stored as an integer, though note that no
check is made to see if double value is within the range of an integer
value, but the constant will not be used as an integer if it is not
within the range of an integer (there is a check when a double constant
is changed to an integer constant to prevent this).

When the integer value is converted from a double, it should have been
rounded up (positive) or down (negative), in other words, away from
zero.  This test initially failed because this was not happening.

The test was made to pass by changing the create double code function of
the constant number helper class to round the double value, by calling
the standard round function returning a long integer, and then static
casting to a 32-bit integer.  Through this may be a narrowing
conversion, it doesn't matter since the integer value will not be used
if the double value is not in the range of a 32-bit integer.