const-num: corrected limit checks on integer equivalent of double value
TESTS: verify that the limit checks take into account rounding for the
integer equivalent value of a double constant.

Two of these tests, constants with a ".499" fractional part, initially
failed because the within integer range function was checking against
the minimum and maximum 32-bit integer values, so any number with a
non-zero fractional part triggered an out of range error.  If the
fractional part is less than 0.5, the number would round towards zero
and would be within the limit.

The tests were made to pass by modifying the within integer range
function to take the minimum or maximum 32-bit integer limit, static
cast it to a double and then add or subtract 0.5 and put these into
local upper and lower out of range variables to which the input value is
compared to.  Since this functionality is not applicable for 64-bit
integer input values, the template function was separated into two
functions, one for double types and one for 64-bit integer types.