const-str: parse initial double quote of a string constant
TEST: verify that the opening double quote of a string constant is
parsed.

The test checks that after calling the new compile string constant
function parses past the opening double quote.

To get the code to compile, the a new compile string constant function
was added to the compiler class, which has no arguments and no return
value.  The test did not pass since this function didn't do anything.

The get the test to pass, code was added to the new compile string
constant function that checks if the next character is a double quote,
and if it is, it gets the character.