const-str: parse to the ending double quote of a string constant
TEST: verify that the string constant is parsed to the end of the
string.

The compile string constant function was rearrange a bit where the check
for the opening double quote was reversed and if the next character is
not a double quote the function returns, otherwise the double quote is
consumed.

To get the test to pass, a while loop was added, which continues
consuming characters until the end of the input, which is sufficient to
get the test to pass since the test string constant is the only thing in
the input string.