const-str: allow two consecutive double quotes to be part of constant
TEST: verify that when two double quotes appear next to each other that
they are included in the string constant.

This test initially failed because the compile string constant function
terminated when finding a double quote.

The test was made to pass by adding a check after seeing the first
double quote, that the function only returns if the following character
is not another double quote.