const-str: added data type return value
TESTS: verify that the correct data type is returned to whether a string
was parsed of not.

The first test checked for a null return data type (for a non-string).
In order to make this compile, the return type of the compile string
constant function was changed from void to the data type enumeration,
the null data type was added to return statements in the function (the
second return was changed to a break and the return was moved to the
end).

The second test checked for a string return data type (for a string
constant).  A new string data type enumerator needed to be added to get
this to compile.  The second return statement was changed to return the
string data type to get this test to pass.