const-str: added constant string code to program
TEST: verify that a constant string code is added to the program code
with its operand.

The first of two checks in this test verifies that two program words
have been added.  This check prevents undefined behavior on the second
check that verifies the first word contains the constant string code
value.

To get the test to compile and pass required a new constant string code
instance to be added (with null function pointers), and the compile
string constant function adding the code value of the constant string
code and a dummy operand of zero to the code line.