const-str: added dictionary to hold string constants
TEST: verify that each constant has a unique operand (index) value.

This test initially failed because the operand of the constant string
code was hard-coded to a zero.  To get this test to pass, a constant
string dictionary needed to be added to hold the string constants.

A new constant string dictionary class was implemented, which is derived
from the dictionary base class.  It has a single vector member to hold
the string values where each member is a standard unique pointer to a
standard string.  A string pointer was chosen since these will be put on
the execution stack, it is desirable for a string to be a small as
possible and a pointer is the same size as a double variable (or an
integer variable on a 32-bit platform).

This new class as a single add member function taking a string as an
argument.  This function adds the string to the dictionary.  If it
didn't exist in the dictionary already, it is added to the string values
vector.  The operand (index) of the string is returned.

An instance to a constant string dictionary was added to the program
unit class along with an access member to return a reference to this
dictionary.

The compile string constant function needed to be modified to create a
string of the string constant as it was only previously parsing the
string but not doing anything with the characters of the string.  The
string parsed is added to the constant string dictionary and the operand
returned is added to the code line instance of the hard-coded zero.