const-str: handle empty string constants
TEST: verify that an empty string is parsed corrected and stored in the
constant string dictionary.

This test initially failed because the code in the compile string
constant function was incorrectly adding the second quote character of
an empty string.

The statements of the compile string constant were rearranged using a
temporary variable to hold the character from the input stream until
being adding it to the string.

For the test, a new constant string dictionary access function for a
constant program instance was added to the program unit class.  This
function will be used later for recreating constant strings.