program: encapsulated constant number and string dictionaries
The program unit class contained access functions that return references
to the dictionary instances.  Callers used these functions to call
dictionary member functions with a chain of calls.  This was not good
object oriented design.

The functions were replaced with an add constant number, an is constant
convertible to integer, a get constant number, an add constant string,
and a get constant string functions.