data-type: encapsulated enumeration inside a class
This was done so that checks that will need to be added and putting
these checks inside a class the holding the data type value is the most
appropriate place.  For example, there will be a need to check if the
data type is numeric, either double or integer.

The data type enumeration was moved into a new class with the same name,
and the enumeration class name was renamed.  The enumeration was made
private.  There are static creation functions that return data type
instances for a specific data type (double, integer and string).
There is a default constructor (sets to the null data type).  There is
an explicit boolean conversion operator for testing if the data type
doesn't contain the null data type.  And there are is functions for
testing for a particular type (double, integer and string).

There is a private constructor for initializing an instance for a
specific data type that is used by the static creation function.  Since
the enumeration is private, the enumerators couldn't be used outside of
this class, therefore this constructor can't be used outside either.

All uses of the data type enumeration were replaced with the appropriate
member functions of the new data type enumeration.  Use of the null data
type were replaced with using either the default constructor, or the
explicit boolean conversion operator.  The compile numeric expression
and compile parentheses functions were given a default value for the
data type as the default constructor (in other words, a null data type).

For some strange reason, the coverage test shows an uncovered line in
the compiler compile expression taking an expected data type, the last
line in the function (the implied return at the closing quote).  It
appears this is due to the compile generating a jump to the expression
compiler compile expression function instead of a call.  Not sure why
this was not the case before this data type change (may also be due to
the change in GCC version from 4.9.3 to 4.9.4).

OTHER: the special string literals used in the tests for creating string
constants were simplified a bit - the asterisks characters were
unnecessary since there were no double quotes followed by an open
parenthesis character inside any of the strings.  These extra
deliminator characters are only needed if these characters are found
inside the string.