not: made sure that there is an operand after not operator
TEST: verify that an error is produced if there is not a valid numeric
expression following a NOT operator.

This check was previously missed.  This test initially failed with a
segmentation fault because the compiler entered the not code into the
program, but without any operand, and when the not code executed, the
execution stack was empty causing the crash.

To get this test to pass, a check for no data type was added to the
compile not operand function after calling the compile not function.