compiler: added ability to compile string expressions
TEST: verify that the compile returns a string data type when compiling
a string constant.

This test initially failed because the compile expression function,
which uses the expression compiler helper class, only handled numeric
expressions, so with the expression starting with a string constant (a
double quote) was not understood causing a null data type to be
returned.

The test was made to pass by changing the compile numeric operand
function to call the compile string constant if the operand wasn't a
parenthetical expression, not operator, function, or a numeric constant.
The name "compile numeric operand" no longer was accurate and so was
renamed to just the "compile operand" function.  The string data type
then trickles up to the top compile expression allowing the string data
type to be returned.