const-str: produce error if operand to a math function
TEST: verify that an error is produced if a string constant is used with
a math operator.

This test initially failed because an error was expected but none was
thrown because the compile operator function in the expression compiler
only checked if the data type was not a null, it did not validate that
the data was numeric before proceeding.

To make this test pass, a check was added after getting valid operand
codes, that if the data type is not numeric, then an expected numeric
expression error is thrown.  The check for non-numeric is a check for
neither a double or an integer data type.  An is numeric function was
added to the new data type class to check if the data type is double or
integer.  An is not numeric function was added as the not of this
function (because it reads easily with 'not' then with the not '!'
operator at the beginning of the if expression where it not very
visible.