const-str: error column and length when operand to a math operator
TEST: verify that the error column and length are correct when a string
constant is an operand to a math operator.

The test string puts the string constant away from the beginning of the
string (so that the error column is not zero).  This test initially
failed because the error column was set to zero and the length was set
to the default of one.

To get this test to pass, code was added to the generic compile operator
function was to record the column of the operand before compiling the
operand, and afterward calculating the operand length from the column
after the operand minus the column of the operand.  These are passed
to the expected numeric expression error instance that is thrown if the
operand is not numeric.  This required a length argument on this error
class, which was given a default value of one and is passed to the base
compile error class.