const-str: produce error if right operand of a math operand
TEST: verify that an error is produced if a string constant is used on
the right hand side of a math operator.

The first test (checking for a thrown error) initially failed because
the generic compile operator function on checked if the right hand side
data type was set.  This check was changed to is not numeric.

The second test (checking the column and length) was made to pass the
same as for the left side operand - recording the column before
compiling the operand, and calculating the length afterward.