const-str: produce error if operand of a not operator
TEST: verify that an error is produced if a string constant is used as
the operand of a NOT operator.

The first test (checking for a thrown error) initially failed because
the compile not operand function only checked if the operand data was
set.  This check was changed to is not numeric.

The second test (checking the column and length of the error) was made
to pass the same as the other instance - recording the column before
compiling the operand (an exponential expression), and calculating the
length afterward.