fix: corrected error lengths not caught
In the process of making the last few changes, only the strings unit
tests were being run (going to fast).  A few of the other unit tests
were failing due to the operand length being calculated to zero.  This
occurred when there was an unrecognizable character was found so the
operand column before and after was the same.

The compile not operand, compile negation, and compile operator (right
hand side) at the checks for a non-numeric data type, if the data type
was null, the length is set to one.  The length calculation was also
moved into the if since it is only needed when an error is thrown.  For
the left hand side, the length needs to be calculated right after the
operand because the getting of the operator changes the column.

The statements are duplicated, so this needs to be refactored.  Also,
the generic compile operator function is getting larger then it should
be and also needs to be refactored.