expression: refactored to simplify, remove duplication and clean up
The compile numeric expression functions were only called from the
compile expression main functions, and the string data type is now being
passed up from below.  Only if a numeric expression was requested did
one of these functions optional convert to a particular data type
(double or integer).  The actual check for a numeric data type was in
the compile expression function.  The other compile numeric expresion
function was just a pass through and did nothing else.

Therefore, both of the compile numeric expression functions were
removed.  The convert functionality for the one was moved into the
compile expression function taking an expected data type.  This function
may need further refactoring once string support is fully implemented.

The next refactoring involved removing code duplication.  There was code
in the compile not operand, compile negation, and compile operator (for
right hand side operand) that was all the same - obtained the current
column (for the operand), compiled the operand, and if the data type of
the operand was not numeric, would calculate the length of the operand
(from the now current column) and throw an "expected numeric expression"
error.  The only difference was the function called to compile the
operand.

Therefore, the new compile numeric operand function was creating by
extracting this code, was given a member function pointer to the
function to use to compile the operand, and has a data type return value.

The final refactorings was to rename the compile operand function
pointers to compile sub-expression to clarify what is actually being
performed (it's not just compiling an operand).  Following suit, the
compile numeric operand function was renamed to compile numeric
expression (again it's not just a numeric operand).  And finally, type
aliases were added for the compile sub expression, get operator codes
(renamed from just get codes), and convert function pointers.  This make
it easier to see that these function arguments are function pointers
(the embedded function pointer syntax hides the name of the argument).
 add type aliases for the