const-str: produce error if argument to multi-type function
TEST: verify that an error is produced if a string constant is used as
an argument to a multi-type function (ABS).

All of the functions implemented so far request a specific data type
(integer or double) except for the multi-type functions (ABS and SGN).
These functions allow any numeric type (there is a code for each data
type).

This test initially failed because any data type was accepted including
string (an error is expected).  A double code was selected if the data
type of the argument is double, otherwise the integer code was selected.

The first test (that an error is throw) was made to pass by adding a
check to the compile expression function to throw an error if a non-null
data type is not numeric.  The second test failed because the error
contained the wrong column and length.

The second test was made to pass by recording the column in the compile
expression function before calling the compile implication function and
calculating the length afterward.

This code was identical to the code in the generic compile operator
function for the left side operand except for the function used to
compile the sub-expression.  To eliminate this duplication, a new
compile sub-expression was added to get the current column, call a
sub-expression function (passed as a function pointer argument), and
calculate the length of the sub-expression.

So that all three values (column, length and data type of the
sub-expression) can be returned, a private sub-expression structure was
added.