const-str: implemented recreation of string constants
TEST: verify that a string constant is recreated correctly.

This test initially failed because there was no recreate function
pointer in the constant string code instance.

The constant string code instance was moved from the compiler source
file to the constant string source file.  A recreate constant string
function was added.  A local string is set to a double quote, a new get
constant string operand function in the recreator class is called to get
the string constant value from the program to add to the local string,
the ending double quote is added, and the string is pushed to the
recreator stack.

The new virtual get constant string operand function of the recreator
class gets the operand of the constant string code, which is passed the
program unit get constant string.  The returning string is returned.