const-str: recreate string constants with embedded double quotes
TEST: verify that a string constant with an embedded double quote is
recreated correctly.

This test initially failed because an embedded double quote was
recreated as a single double quote instead of two double quotes.

The recreate constant string function was modified to iterate the string
returned from the recreator get constant string operand function and add
each character one at a time.  If a character added is a double quote,
a second one is added to the string.