const-str: added compiling of strings to the print command
TEST: verify that a PRINT command with a string constant is compiled
correctly.

To get this test to compile, a new print string code was added (starting
with null function pointers).  The test initially failed because the
compile print function was adding a print integer code after the string
constant.

The get this test to pass, the compile print function was modified to
check if the data type is integer for adding a print integer code,
otherwise a print string code is added to the program.