const-str: implemented execution of string constants
TEST: verify that a string constant is executed correctly.

To get this test to compile, the executer class needed a top string
function.  Though not needed for compilation (but needed for execution),
a standard string pointer member was added to the executer stack item.
The compiler add string constant instruction also needed to be made
public to be accessible from the test (this design probably needs to be
reconsidered).

For the new execute constant string function, the executer class needed
a push constant string.  For this function, the executer class needed
a pointer to the array of string constants.  The constant string
dictionary class stores these as a vector of unique string pointers.
The push constant string function needs to get the actual string pointer
from the unique pointer in the appropriate element.  The stack item
class also needed a new constructor taking a standard string pointer.

A pointer to an array of unique standard string pointers was added to
the executer class and the program unit create executer function (whose
name was misspelled and was renamed) was updated accordingly.  This
required a get string values function be added to the constant string
dictionary, which returns the pointer to the data of the unique standard
string pointer vector.