print: execute a print statement with a string constant
TEST: verify that a PRINT statement with a string constant is executed
correctly.

To get this test to pass, an execute print item string function was
added that pops the top string pointer from the execution stack, outputs
its contents to the output stream, and pops the value from the stack.  A
pointer to this function was added to the print string item code
instance replacing the null function pointer.