less-than: implemented parsing with two string operands
TEST: verify that a less than expression with two string operands are
parsed to the end of the expression.

The test initially failed because a the less-than (relation) operator
expected a numeric expression as an operand.  Comparison (relation and
equality) operators allow string operands in addition to numeric
operators, and the generic compile operator function only handles
operators with numeric operands.

A new compile comparison operator function was added with similar
functionality as the generic operator function with some differences.
This function does not contain a convert function pointer argument as
these operators don't need conversions added.  The data type of the
left side operand does not need to be checked if numeric since a string
is also allowed.  The compile sub-expression function pointer can be
called directory instead of calling the compile numeric expression,
which also checks for numeric only expressions.

The compile relation function was changed from calling the generic
compile operator function to calling this new compile comparison
operator function.  These changes were sufficient to get this test to
pass.