less-than: implemented compiling of operator with two string operands
TEST: verify that a less than expression with two string operands are
compiled correctly.

To get this test to compile, a less than string-string code instance
was added.  This required a new string-string operator type enumerator.
The code instance was given null function pointers.

To get this test to pass, a number of extra changes were made, changes
that were no strictly needed, but will be needed, and so were added,
otherwise some backward compatible code were be needed temporarily that
would have to be removed later (so the former changes were made).

To this end, the comparison operator codes class was modified to also
contain a reference to a string-string code, and the constructor was
updated to initialize it.  This required this class to be broken out of
the class hierarchy where it now is derived from the base operator codes
instead of a the numeric operator codes.  Its select function was fully
implemented including check for string data types.  The code values
function was also implemented.

The changes to the comparison operator codes class required comparison
string-string code instances for all of the relation and equality
operators and the relation and equality codes instances updated.  All
code instances were given null function pointers.

The comparison operator codes select function was implemented to throw
an error (an empty expected numeric operator error class) if there is
no code for the data types (for example a string and a number data
type).  (This isn't right for when expecting a string operand, but no
test exists yet for that condition.)

The full expected numeric expression error class wasn't used because
that class requires a column and length, which is obtained from the
compiler instance, something not available to the select function.  This
error is caught in the compile comparison operator function, and the
expected number expression error is rethrown with the column and length
of the error.  Since a null data type would also be caught by the checks
if the select function, the check for a null data type was removed from
the compile comparison operator function.